requirements.txt
# Jupyter notebook
jupyter

# Data Processing & Management
datasets
pandas
numpy

# Visualize result
matplotlib

# Core Deep Learning
transformers

# Information Retrieval & Embedding
sentence-transformers
faiss-cpu
rank_bm25
py_vncorenlp

# Utilities & Monitoring
tqdm

# Tokenize
underthesea
py_vncorenlp
spacy

# Word Embedding
gensim

# Format Python string to dictionaries
conllu